profiles/workstation: Add xpm USE flag to dev-libs/efl, prevents E restart
This seems to be the cause of issues with E restarting on particular
icons, that happen to be xpm format. Not all xpm images trigger the
issue. This is basically what was reported to upstream. Need the
combination of EFL built without support for xpm, and running E22. E21
did not trigger the issue, it is new with E22. It seems to be a bug in
E22 not EFL.

https://phab.enlightenment.org/T6146